:sparkles: feat: update archives

+++ JavaBankSystem/src/main/java/br/com/dio/model/Investment.java
@@ -1,8 +1,11 @@
 package br.com.dio.model;
 
+import java.math.BigDecimal;
+
+// Representa um produto de investimento disponível no sistema
 public record Investment(
-        long id,
-        long tax,
-        long daysToRescue,
-        long initialFunds
+        long id,              // ID único do investimento
+        long tax,             // Taxa de rendimento (percentual)
+        long daysToRescue,    // Dias mínimos para resgatar
+        BigDecimal initialFunds     // Valor inicial exigido
 ) {}
